Add typings for payment.
