+++ src/operation.ts
@@ -283,5 +283,4 @@ export class Operation extends BaseOperation {
 
 // Attach all imported operations as static methods on the Operation class
 // TS-TODO: move this to BaseOperation
-Operation.payment = ops.payment;
 Operation.setOptions = ops.setOptions;
